Fixed a lot of bugs in the algorithem, it now works as it was inteded, But it does not work.

+++ Firmware/features/StandardDeviation.h
@@ -17,7 +17,7 @@ typedef struct stdDev{
 	float variance;
 	uint16_t * StdSampleList;
 	uint8_t active;
-	uint8_t listIndex;
+	uint16_t listIndex;
 } stdDev;
 
 void StdDev_Reset(stdDev * this);
